pyproject.toml
[tool.poetry]
name = "bot_rain"
version = "0.1.0"
description = ""
authors = ["intmax"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
graia-ariadne = "^0.7.16"
prettytable = ">=2.5.0"
pycryptodome = ">=3.12.0"
qrcode = ">=7.3.1"
Pillow = ">=9.0.0"
SQLAlchemy = ">=1.4.29"
matplotlib = ">=3.5.1"
ujson = ">=5.1.0"
graia-scheduler = ">=0.0.6"
graia-saya = ">=0.0.13"
regex = ">=2021.11.10"
graia-broadcast = ">=0.14.5"
aiohttp = {version = ">=3.7.4.post0", extras = ["speedups"]}
playwright = ">=1.17.2"
aiofile = "^3.7.4"
imageio = "^2.16.1"
lxml = "^4.9.1"
saucenao-api = "^2.4.0"
easyocr = "^1.4.2"
bs4 = "^0.0.1"
html2image = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
